Update spreadsheets.py
Reverting prior commit, as it did not resolve the issue.  Also, it was found that the issue is actually not in rsoxs_scans.  Running load_samplesxlsx casts edge as a list when a list is provided.  However, it is when it is saved into rsoxs_config["bar"] in the rsoxs package that the datatype is altered.